Add configuration element.
The configuration components allows everything to come to life by allowing the theme to be modified.
State related data is manipulated in the element, while view data is manipulated in the component.
This is a working version of the layout manager and theming.
I can start the adding content now, despite still needing some things in the core.