Add canvas triangles background.
I just wanted to mess around with triangles to see if I could get a background I found looked good enough to publish.
Some refactoring will be done to make this a bit more generic.
The manager also now keeps track of window events, and I will probably use RXJS for making that cleaner.
Ultimately the component switching is gonna be completely separated from react so that the actual components can work in a react-y way.