Remove CSS modules type definitions from .gitignore.
None of the available CSS typing generators seem to work and I don't have time to fix them or write my own.
I'll write the typings manually as it's just a matter of copying a few class names.
While it's not ideal, it will have to do until a good generator appears.