Add default project viewer styling.
The original idea was to make the child component send data to the page header.
That way the page header could react to the displayed project.
That will require a different approach to state management and code that I don't have the time to write now.
The current solution is just a collapsible list of projects.
Fix a few styling mistakes that don't need a separate commit.